Fixed Boss Changes committed: 	modified:   include/Mounts/AlphaOneMount.h 	modified:   include/Weapons/AlphaOnePlasma.h 	modified:   include/Weapons/Weapon.h 	modified:   src/Drops/Coins.cpp 	modified:   src/Drops/Health.cpp 	modified:   src/Drops/Shield.cpp 	modified:   src/Enemies/AlphaOne.cpp 	modified:   src/Enemies/BasicEnemy.cpp 	modified:   src/Enemies/CorruptedStarShip.cpp 	modified:   src/Enemies/DotEnemy.cpp 	modified:   src/Enemies/HypnoEnemy.cpp 	modified:   src/Mounts/AlphaOneMount.cpp 	modified:   src/Weapons/AlphaOnePlasma.cpp

+++ include/Mounts/AlphaOneMount.h
@@ -16,7 +16,7 @@ class AlphaOneMount: public Mount {
     void tick(float x, float y, float deltaTime, float directionX, float directionY, float angle, bool isShooting);
 
   protected:   
-    std::vector<float> cycle;
+    std::vector<int> cycle;
     std::vector<float> Nx;
     std::vector<float> Ny;
 };
